requirements.txt
Django==4.2.26
djangorestframework==3.16.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
firebase-admin==6.3.0
firebase-functions==0.1.0
